refactor: move Nexus-related tests to protocol_test for cleanup and modularity
